refactor: move config and incidents under relevant directories
