+++ src/components/Configs/Sidebar/ConfigChanges/Cells/ConfigChangeAgeCell.tsx
@@ -1,6 +1,6 @@
 import { CellContext } from "@tanstack/react-table";
-import { Age } from "../../../../ui/Age";
-import { ConfigChange } from "../../../../api/types/configs";
+import { Age } from "../../../../../ui/Age";
+import { ConfigChange } from "../../../../../api/types/configs";
 
 export default function ConfigChangeAgeCell({
   row,
