+++ src/components/Configs/Sidebar/configs.tsx
@@ -4,19 +4,19 @@ import { useEffect, useState } from "react";
 import { TbTrash, TbTrashOff } from "react-icons/tb";
 import { VscJson } from "react-icons/vsc";
 import ReactTooltip from "react-tooltip";
-import { useComponentConfigRelationshipQuery } from "../../api/query-hooks/useComponentConfigRelationshipQuery";
-import { removeManualComponentConfigRelationship } from "../../api/services/configs";
-import { Badge } from "../Badge";
-import CollapsiblePanel from "../CollapsiblePanel";
-import ConfigLink from "../Configs/ConfigLink/ConfigLink";
-import { ConfirmationPromptDialog } from "../Dialogs/ConfirmationPromptDialog";
-import EmptyState from "../EmptyState";
-import { IconButton } from "../IconButton";
-import TextSkeletonLoader from "../SkeletonLoader/TextSkeletonLoader";
-import { refreshButtonClickedTrigger } from "../SlidingSideBar";
-import Title from "../Title/title";
-import { toastError, toastSuccess } from "../Toast/toast";
-import TopologyConfigsActionsDropdown from "../Topology/Sidebar/Utils/TopologyConfigsActionsDropdown";
+import { useComponentConfigRelationshipQuery } from "../../../api/query-hooks/useComponentConfigRelationshipQuery";
+import { removeManualComponentConfigRelationship } from "../../../api/services/configs";
+import { Badge } from "../../Badge";
+import CollapsiblePanel from "../../CollapsiblePanel";
+import ConfigLink from "../ConfigLink/ConfigLink";
+import { ConfirmationPromptDialog } from "../../Dialogs/ConfirmationPromptDialog";
+import EmptyState from "../../EmptyState";
+import { IconButton } from "../../IconButton";
+import TextSkeletonLoader from "../../SkeletonLoader/TextSkeletonLoader";
+import { refreshButtonClickedTrigger } from "../../SlidingSideBar";
+import Title from "../../Title/title";
+import { toastError, toastSuccess } from "../../Toast/toast";
+import TopologyConfigsActionsDropdown from "../../Topology/Sidebar/Utils/TopologyConfigsActionsDropdown";
 
 type Props = {
   topologyId?: string;
