+++ src/components/Incidents/Sidebars/incidents.tsx
@@ -3,19 +3,19 @@ import { useAtom } from "jotai";
 import { useEffect, useMemo, useState } from "react";
 import { ImLifebuoy } from "react-icons/im";
 import { Link } from "react-router-dom";
-import { getIncidentsBy } from "../../api/services/incident";
-import { useFeatureFlagsContext } from "../../context/FeatureFlagsContext";
-import { features } from "../../services/permissions/features";
-import PillBadge from "../Badge/PillBadge";
-import CollapsiblePanel from "../CollapsiblePanel";
-import { DetailsTable } from "../DetailsTable/DetailsTable";
-import { IncidentStatusTag } from "../IncidentStatusTag";
-import { typeItems } from "../Incidents/data";
-import IncidentsFilterBar, { IncidentFilter } from "../IncidentsFilterBar";
-import { refreshButtonClickedTrigger } from "../SlidingSideBar";
-import Title from "../Title/title";
-import { Age } from "../../ui/Age";
-import { IncidentTypeIcon } from "../incidentTypeTag";
+import { getIncidentsBy } from "../../../api/services/incident";
+import { useFeatureFlagsContext } from "../../../context/FeatureFlagsContext";
+import { features } from "../../../services/permissions/features";
+import PillBadge from "../../Badge/PillBadge";
+import CollapsiblePanel from "../../CollapsiblePanel";
+import { DetailsTable } from "../../DetailsTable/DetailsTable";
+import { IncidentStatusTag } from "../../IncidentStatusTag";
+import { typeItems } from "../data";
+import IncidentsFilterBar, { IncidentFilter } from "../../IncidentsFilterBar";
+import { refreshButtonClickedTrigger } from "../../SlidingSideBar";
+import Title from "../../Title/title";
+import { Age } from "../../../ui/Age";
+import { IncidentTypeIcon } from "../../incidentTypeTag";
 
 type Props = {
   topologyId?: string;
